+++ packages/opencensus-exporter-instana/src/instana.ts
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-import {Exporter, ExporterBuffer, ExporterConfig, RootSpan, Span} from '@opencensus/core';
+import {Exporter, ExporterBuffer, ExporterConfig, RootSpan, Span, SpanKind} from '@opencensus/core';
 import {logger, Logger} from '@opencensus/core';
 import {request} from 'http';
 
 const spanKindTranslation: {[k: string]: string} = {
   CLIENT: 'EXIT',
   SERVER: 'ENTRY',
-  LOCAL: 'INTERMEDIATE',
+  UNSPECIFIED: 'INTERMEDIATE',
 };
 
 type InstanaSpan = {
@@ -109,7 +109,7 @@ export class InstanaTraceExporter implements Exporter {
       timestamp: span.startTime.getTime(),
       duration: span.duration,
       name: span.name,
-      type: spanKindTranslation[span.kind] || span.kind,
+      type: spanKindTranslation[span.kind],
       error: span.status != null && span.status.code !== 0,
       data: Object.keys(span.attributes)
                 .reduce(